Revamp async support
* Use callbacks for papillon-native async support (no dependencies, easily
  adaptable).
* Add multi-arity Chrysalis protocol for optimized sync and async resolution,
  allowing the sync interface to stand alone even in ClojureScript.
* Extract interceptor-to-interceptor movement operation (a pure function) from
  interceptor execution function by representing the _stage_ of the interceptor
  in the context.
* Simplify the interceptor tagging/naming mechanism.
* Add unifying async test helper for Clojure and ClojureScript.

Where to go from here:
**A. We could simplify the dual arities of Chrysalis by expressing the sync API
as a thin wrapper around an async-only Chrysalis API.  On the plus side, the
Chrysalis protocol is simplified.  On the downside, the sync API cannot work
in single-threaded environments (ClojureScript), and the callback-only variant
of Chrysalis is an impedance mismatch (unnatural) for sync-only work in Clojure.
**B. Now that core.async is no longer a dependency of papillon proper, we
could double down on opt-in, batteries-included support for promesa, manifold,
etc.  That work could also serve to demonstrate the ease of implementing emerge
for various paradigms.  This allows all manner of interceptors to work seam-
lessly together.
**C. Beyond extending Chrysalis to alternate async models, we could demonstrate
the ease with which one can wrap the outer-most callback approach of papillon to
various alternative paradims.  It would be easy to write transition wrappers
that convert the callback API to core.async (Clojure and ClojureScript; done, in
a test); JavaScript promises (also done in a test) and possibly clojure.core
promises/futures/delays; promesa promises (clojure and clojurescript) and
and manifolds.